Adapt 'useDebounce' to TypeScript
